Split style.py into a style/ package (2.0 PR 4, Workstream G)
Pure, behavior-preserving move of the ~5.9k-line style.py into a real
style/ package, per the design pass in development/2_0_style_split_design.md:

  style/
    theme.py        # Colors, ThemeDefinition
    builders_tk.py  # StyleBuilderTK
    builders_ttk.py # StyleBuilderTTK (the bulk)
    engine.py       # Style
    bootstyle.py    # Keywords, Bootstyle, BootMixin/AutoStyleMixin,
                    #   bootify/apply_bootstyle/enable_global_api
    __init__.py     # re-exports the full public surface

Strict downward import layering (theme -> builders_tk -> builders_ttk ->
engine -> bootstyle); the only cycles are 6 function-local back-edges
(Colors->Style, both builders' __init__->Style, engine->Bootstyle x2). No
import-time cycle; each submodule imports standalone.

ttkbootstrap.style stays a valid public path, so every existing
`from ttkbootstrap.style import ...` resolves with no shim, and
`import ttkbootstrap` stays warning-free. Submodule paths are
implementation detail (treat like internal/).

Pruned genuinely-unused imports (engine: tk, ImageTk, Colors; bootstyle:
tk, constants *), verified via pyflakes. New tests/test_style_package.py
adds structural guards (public surface, legacy imports, standalone-submodule
cycle guard, consumer imports).

Python 3.14 / PEP 649 note: lazy annotations masked a missing
annotation-only import (ThemeDefinition in builders_tk) that imported clean
but NameErrors on annotation access. Added an annotation force-evaluation
sweep to the verification; reuse it for the E/D code moves.

Suite: 61 passed.
