+++ src/ttkbootstrap/style/builders_ttk.py
@@ -1,1372 +1,19 @@
-"""Style engine and theme management for ttkbootstrap.
+"""TTK widget style builders for ttkbootstrap.
 
-This module provides the core styling system for ttkbootstrap, managing themes,
-colors, and widget styles. It includes the Style class for applying and managing
-Bootstrap-inspired themes, along with color manipulation utilities.
-
-Classes:
-    Colors: Defines color scheme for themes with manipulation methods
-    ThemeDefinition: Encapsulates complete theme configuration
-    Style: Main style engine for managing themes and widget styles
-    StylerTTK: TTK widget styler
-    StylerTK: Legacy tk widget styler
-    Bootstyle: API for applying bootstyles to widgets
-
-Key Features:
-    - 40+ built-in Bootstrap-inspired themes
-    - Dynamic theme switching at runtime
-    - Custom theme creation and management
-    - Widget-specific style boosting (outline, link, toggle, etc.)
-    - Color utilities for HSV manipulation
-    - Version-stamped theme walk for theme change repaints
-    - Support for user-defined themes
-
-Color Keywords:
-    - primary, secondary, success, info, warning, danger, light, dark
-
-Style Types:
-    - outline, link, toggle, inverse, striped, toolbutton, square
-
-Example:
-    ```python
-    import ttkbootstrap as ttk
-    from ttkbootstrap.constants import *
-
-    # Create window with theme
-    app = ttk.Window(themename="darkly")
-
-    # Access style instance
-    style = ttk.Style.get_instance()
-
-    # Get theme colors
-    primary_color = style.colors.primary
-    bg_color = style.colors.bg
-
-    # Change theme at runtime
-    style.theme_use("flatly")
-
-    # Create styled widgets
-    btn = ttk.Button(app, text="Success", bootstyle="success")
-    btn_outline = ttk.Button(app, text="Danger", bootstyle="danger-outline")
-
-    app.mainloop()
-    ```
+`StyleBuilderTTK` holds the ``create_*_style`` methods that build every ttk
+style (and its image assets via the content-addressed cache). The largest module
+in the package. Split out of the monolithic `style.py` in 2.0.
 """
-import colorsys
-import json
-import re
 import tkinter as tk
 from math import ceil
-from tkinter import TclError, font, ttk
-from typing import Any
+from tkinter import font, ttk
 
-from PIL import Image, ImageColor, ImageDraw, ImageFont, ImageTk
+from PIL import Image, ImageDraw, ImageFont, ImageTk
 from PIL.Image import Resampling, Transpose
 
-from ttkbootstrap import colorutils
-from ttkbootstrap.internal import utility as util
 from ttkbootstrap.constants import *
-from ttkbootstrap.themes.standard import STANDARD_THEMES
-
-try:
-    # prevent app from failing if user.py gets corrupted
-    from ttkbootstrap.themes.user import USER_THEMES
-except (ImportError, ModuleNotFoundError):
-    USER_THEMES = {}
-
-
-class Colors:
-    """A class that defines the color scheme for a theme as well as
-    provides several static methods for manipulating colors.
-
-    A `Colors` object is attached to a `ThemeDefinition` and can also
-    be accessed through the `Style.colors` property for the
-    current theme.
-
-    Examples:
-
-        ```python
-        style = Style()
-
-        # dot-notation
-        style.colors.primary
-
-        # get method
-        style.colors.get('primary')
-        ```
-
-        This class is an iterator, so you can iterate over the main
-        style color labels (primary, secondary, success, info, warning,
-        danger):
-
-        ```python
-        for color_label in style.colors:
-            color = style.colors.get(color_label)
-            print(color_label, color)
-        ```
-
-        If, for some reason, you need to iterate over all theme color
-        labels, then you can use the `Colors.label_iter` method. This
-        will include all theme colors.
-
-        ```python
-        for color_label in style.colors.label_iter():
-            color = Colors.get(color_label)
-            print(color_label, color)
-        ```
-
-        If you want to adjust the hsv values of an existing color by a
-        specific percentage (delta), you can use the `Colors.update_hsv`
-        method, which is static. In the example below, the "value delta"
-        or `vd` is increased by 15%, which will lighten the color:
-
-        ```python
-        Colors.update_hsv("#9954bb", vd=0.15)
-        ```
-    """
-
-    def __init__(
-            self,
-            primary,
-            secondary,
-            success,
-            info,
-            warning,
-            danger,
-            light,
-            dark,
-            bg,
-            fg,
-            selectbg,
-            selectfg,
-            border,
-            inputfg,
-            inputbg,
-            active,
-    ):
-        """
-        Parameters:
-
-            primary (str):
-                The primary theme color; used by default for all widgets.
-
-            secondary (str):
-                An accent color; commonly of a `grey` hue.
-
-            success (str):
-                An accent color; commonly of a `green` hue.
-
-            info (str):
-                An accent color; commonly of a `blue` hue.
-
-            warning (str):
-                An accent color; commonly of an `orange` hue.
-
-            danger (str):
-                An accent color; commonly of a `red` hue.
-
-            light (str):
-                An accent color.
-
-            dark (str):
-                An accent color.
-
-            bg (str):
-                Background color.
-
-            fg (str):
-                Default text color.
-
-            selectfg (str):
-                The color of selected text.
-
-            selectbg (str):
-                The background color of selected text.
-
-            border (str):
-                The color used for widget borders.
-
-            inputfg (str):
-                The text color for input widgets.
-
-            inputbg (str):
-                The text background color for input widgets.
-
-            active (str):
-                An accent color.
-        """
-        self.primary = primary
-        self.secondary = secondary
-        self.success = success
-        self.info = info
-        self.warning = warning
-        self.danger = danger
-        self.light = light
-        self.dark = dark
-        self.bg = bg
-        self.fg = fg
-        self.selectbg = selectbg
-        self.selectfg = selectfg
-        self.border = border
-        self.inputfg = inputfg
-        self.inputbg = inputbg
-        self.active = active
-
-    @staticmethod
-    def make_transparent(alpha, foreground, background='#ffffff'):
-        """Simulate color transparency.
-        
-        Parameters:
-
-            alpha (float):
-                The amount of transparency; a number between 0 and 1.
-
-            foreground (str):
-                The foreground color.
-
-            background (str):
-                The background color.
-
-        Returns:
-
-            str:
-                A hexadecimal color representing the "transparent" 
-                version of the foreground color against the background 
-                color.
-        """
-        fg = ImageColor.getrgb(foreground)
-        bg = ImageColor.getrgb(background)
-        rgb_float = [alpha * c1 + (1 - alpha) * c2 for (c1, c2) in zip(fg, bg)]
-        rgb_int = [int(x) for x in rgb_float]
-        return '#{:02x}{:02x}{:02x}'.format(*rgb_int)
-
-    @staticmethod
-    def rgb_to_hsv(r, g, b):
-        """Convert an rgb to hsv color value.
-
-        Parameters:
-            r (float):
-                red
-            g (float):
-                green
-            b (float):
-                blue
-
-        Returns:
-            tuple[float, float, float]: The hsv color value.
-        """
-        return colorsys.rgb_to_hsv(r, g, b)
-
-    def get_luminance(self, color):
-        """Calculate the luminance of a color.
-
-        Parameters:
-            color (str):
-                A hexadecimal color value.
-        Returns:
-            float:
-                The luminance value of the color.
-        """
-        r, g, b = self.hex_to_rgb(color)
-
-        # Convert RGB to linear RGB
-        r = self._get_luminance_value(r)
-        g = self._get_luminance_value(g)
-        b = self._get_luminance_value(b)
-
-        # Calculate luminance using the WCAG formula
-        return 0.2126 * r + 0.7152 * g + 0.0722 * b
-
-    def _get_luminance_value(self, value):
-        if value <= 0.03928:
-            return value / 12.92
-        else:
-            return ((value + 0.055) / 1.055) ** 2.4
-
-    def get_contrast_ration(self, lum1, lum2):
-        """Calculate the contrast ratio between two luminance values.
-
-        Parameters:
-            lum1 (float):
-                The first luminance value.
-            lum2 (float):
-                The second luminance value.
-
-        Returns:
-            float:
-                The contrast ratio.
-        """
-        if lum1 > lum2:
-            return (lum1 + 0.05) / (lum2 + 0.05)
-        else:
-            return (lum2 + 0.05) / (lum1 + 0.05)
-
-    def get_foreground(self, color_label):
-        """Return the appropriate foreground color for the specified
-        color_label.
-
-        Parameters:
-            color_label (str):
-                A color label corresponding to a class property
-
-        Returns:
-            str:
-                A hexadecimal color value for the foreground color.
-
-        Raises:
-            TypeError: If the color_label is not a valid color property.
-        """
-        if color_label == LIGHT:
-            return self.dark
-        elif color_label == DARK:
-            return self.light
-
-        if not Style().dynamic_foreground:
-            return self.selectfg
-
-        # dynamic foreground selection
-        contrast_with_fg = self.get_contrast_ration(
-            self.get_luminance(self.get(color_label)), self.get_luminance(self.fg)
-        )
-        contrast_with_selectfg = self.get_contrast_ration(
-            self.get_luminance(self.get(color_label)), self.get_luminance(self.selectfg)
-        )
-
-        if contrast_with_fg > contrast_with_selectfg:
-            return self.fg
-        return self.selectfg
-
-    def get(self, color_label: str):
-        """Lookup a color value from the color name
-
-        Parameters:
-
-            color_label (str):
-                A color label corresponding to a class propery
-
-        Returns:
-
-            str:
-                A hexadecimal color value.
-        """
-        return self.__dict__.get(color_label)
-
-    def set(self, color_label: str, color_value: str):
-        """Set a color property value. This does not update any existing
-        widgets. Can also be used to create on-demand color properties
-        that can be used in your program after creation.
-
-        Parameters:
-
-            color_label (str):
-                The name of the color to be set (key)
-
-            color_value (str):
-                A hexadecimal color value
-        """
-        self.__dict__[color_label] = color_value
-
-    def __iter__(self):
-        return iter(
-            [
-                "primary",
-                "secondary",
-                "success",
-                "info",
-                "warning",
-                "danger",
-                "light",
-                "dark",
-            ]
-        )
-
-    def __repr__(self):
-        out = tuple(zip(self.__dict__.keys(), self.__dict__.values()))
-        return str(out)
-
-    @staticmethod
-    def label_iter():
-        """Iterate over all color label properties in the Color class
-
-        Returns:
-
-            iter:
-                An iterator for color label names
-        """
-        return iter(
-            [
-                "primary",
-                "secondary",
-                "success",
-                "info",
-                "warning",
-                "danger",
-                "light",
-                "dark",
-                "bg",
-                "fg",
-                "selectbg",
-                "selectfg",
-                "border",
-                "inputfg",
-                "inputbg",
-                "active",
-            ]
-        )
-
-    @staticmethod
-    def hex_to_rgb(color: str):
-        """Convert hexadecimal color to rgb color value
-
-        Parameters:
-
-            color (str):
-                A hexadecimal color value
-
-        Returns:
-
-            tuple[int, int, int]:
-                An rgb color value.
-        """
-        r, g, b = colorutils.color_to_rgb(color)
-        return r / 255, g / 255, b / 255
-
-    @staticmethod
-    def rgb_to_hex(r: int, g: int, b: int):
-        """Convert rgb to hexadecimal color value
-
-        Parameters:
-
-            r (int):
-                red
-
-            g (int):
-                green
-
-            b (int):
-                blue
-
-        Returns:
-
-            str:
-                A hexadecimal color value
-        """
-        r_ = int(r * 255)
-        g_ = int(g * 255)
-        b_ = int(b * 255)
-        return colorutils.color_to_hex((r_, g_, b_))
-
-    @staticmethod
-    def update_hsv(color, hd=0, sd=0, vd=0):
-        """Modify the hue, saturation, and/or value of a given hex
-        color value by specifying the _delta_.
-
-        Parameters:
-
-            color (str):
-                A hexadecimal color value to adjust.
-
-            hd (float):
-                % change in hue, _hue delta_.
-
-            sd (float):
-                % change in saturation, _saturation delta_.
-
-            vd (float):
-                % change in value, _value delta_.
-
-        Returns:
-
-            str:
-                The resulting hexadecimal color value
-        """
-        r, g, b = Colors.hex_to_rgb(color)
-        h, s, v = colorsys.rgb_to_hsv(r, g, b)
-
-        # hue
-        if h * (1 + hd) > 1:
-            h = 1
-        elif h * (1 + hd) < 0:
-            h = 0
-        else:
-            h *= 1 + hd
-
-        # saturation
-        if s * (1 + sd) > 1:
-            s = 1
-        elif s * (1 + sd) < 0:
-            s = 0
-        else:
-            s *= 1 + sd
-
-        # value
-        if v * (1 + vd) > 1:
-            v = 0.95
-        elif v * (1 + vd) < 0.05:
-            v = 0.05
-        else:
-            v *= 1 + vd
-
-        r, g, b = colorsys.hsv_to_rgb(h, s, v)
-        return Colors.rgb_to_hex(r, g, b)
-
-
-class ThemeDefinition:
-    """Encapsulates the name, color palette, and metadata for a
-    ttkbootstrap theme.
-
-    A ThemeDefinition is a lightweight container that pairs a theme name
-    with its Colors object and whether it is a light or dark theme. The
-    Style engine consumes ThemeDefinition instances to build widget
-    styles and images for the active theme.
-    """
-
-    def __init__(self, name, colors, themetype=LIGHT):
-        """
-        Parameters:
-
-            name (str):
-                The name of the theme.
-
-            colors (Colors or dict):
-                A Colors instance or a dict of color values.
-
-            themetype (str):
-                Specifies whether the theme is **light** or **dark**.
-        """
-        self.name = name
-        self.colors = colors if isinstance(colors, Colors) else Colors(**colors)
-        self.type = themetype
-
-    def __repr__(self):
-        return " ".join(
-            [
-                f"name={self.name},",
-                f"type={self.type},",
-                f"colors={self.colors}",
-            ]
-        )
-
-
-class Style(ttk.Style):
-    """A singleton class for creating and managing the application
-    theme and widget styles.
-
-    This class is meant to be a drop-in replacement for `ttk.Style` and
-    inherits all of it's methods and properties. However, in
-    ttkbootstrap, this class is implemented as a singleton. Subclassing
-    is not recommended and may have unintended consequences.
-
-    Examples:
-
-        ```python
-        # instantiate the style with default theme
-        style = Style()
-
-        # instantiate the style with another theme
-        style = Style(theme='superhero')
-
-        # check all available themes
-        for theme in style.theme_names():
-            print(theme)
-        ```
-
-    See the [Python documentation](https://docs.python.org/3/library/tkinter.ttk.html#tkinter.ttk.Style)
-    on this class for more details.
-    """
-
-    instance = None
-
-    def __new__(cls, theme=None):
-        if Style.instance is None:
-            return object.__new__(cls)
-        else:
-            return Style.instance
-
-    def __init__(self, theme=DEFAULT_THEME):
-        """
-        Parameters:
-
-            theme (str):
-                The name of the theme to use when styling the widget.
-        """
-        if Style.instance is not None:
-            if theme != DEFAULT_THEME:
-                Style.instance.theme_use(theme)
-            return
-        self._theme_objects = {}
-        self._theme_definitions = {}
-        self._style_registry = set()  # all styles used
-        self._theme_styles = {}  # styles used in theme
-        self._theme_names = set()
-        # Monotonic counter bumped on every theme switch. The theme walk
-        # repaints (and restamps) only widgets stamped older than this, so a
-        # widget is repainted at most once per switch. Replaces the old
-        # Publisher broadcast.
-        self._theme_version = 0
-        # Content-addressed image cache shared across themes/builders. Keyed on
-        # the pixel-determining inputs of each asset (resolved colors, scaled
-        # size, geometry), NOT the theme name, so cross-theme-identical assets
-        # dedupe and are rendered once. Replaces the per-builder theme_images
-        # dicts that pinned a fresh PhotoImage per theme (the image leak).
-        self._image_cache = {}
-        self._load_themes()
-        self._dynamic_foreground = False
-        super().__init__()
-
-        Style.instance = self
-        self.theme_use(theme)
-
-        # apply localization
-        from ttkbootstrap import localization
-        localization.initialize_localities()
-
-    @property
-    def colors(self):
-        """An object that contains the colors used for the current
-        theme.
-
-        Returns:
-
-            Colors:
-                The colors object for the current theme.
-        """
-        theme = self.theme.name
-        if theme in list(self._theme_names):
-            definition = self._theme_definitions.get(theme)
-            if not definition:
-                return []  # TODO refactor this
-            else:
-                return definition.colors
-        else:
-            return []  # TODO refactor this
-
-    def configure(self, style, query_opt: Any = None, **kw):
-        if query_opt:
-            return super().configure(style, query_opt=query_opt, **kw)
-
-        if not self.style_exists_in_theme(style):
-            ttkstyle = Bootstyle.update_ttk_widget_style(None, style)
-        else:
-            ttkstyle = style
-
-        if ttkstyle == style:
-            # configure an existing ttkbootrap theme
-            return super().configure(style, query_opt=query_opt, **kw)
-        else:
-            # subclass a ttkbootstrap theme
-            result = super().configure(style, query_opt=query_opt, **kw)
-            self._register_ttkstyle(style)
-            return result
-
-    def theme_names(self):
-        """Return a list of all ttkbootstrap themes.
-
-        Returns:
-
-            list[str, ...]:
-                A list of theme names.
-        """
-        return list(self._theme_definitions.keys())
-
-    def register_theme(self, definition):
-        """Register a theme definition for use by the `Style`
-        object. This makes the definition and name available at
-        run-time so that the assets and styles can be created when
-        needed.
-
-        Parameters:
-
-            definition (ThemeDefinition):
-                A `ThemeDefinition` object.
-        """
-        theme = definition.name
-        self._theme_names.add(theme)
-        self._theme_definitions[theme] = definition
-        self._theme_styles[theme] = set()
-
-    def theme_use(self, themename=None):
-        """Changes the theme used in rendering the application widgets.
-
-        If themename is None, returns the theme in use, otherwise, set
-        the current theme to themename, refreshes all widgets and emits
-        a ``<<ThemeChanged>>`` event.
-
-        Only use this method if you are changing the theme *during*
-        runtime. Otherwise, pass the theme name into the Style
-        constructor to instantiate the style with a theme.
-
-        Parameters:
-
-            themename (str):
-                The name of the theme to apply when creating new widgets
-
-        Returns:
-
-            Union[str, None]:
-                The name of the current theme if `themename` is None
-                otherwise, `None`.
-        """
-        if not themename:
-            # return current theme
-            return super().theme_use()
-
-        # change to an existing theme
-        existing_themes = super().theme_names()
-        if themename in existing_themes:
-            self.theme = self._theme_definitions.get(themename)
-            super().theme_use(themename)
-        # setup a new theme
-        elif themename in self._theme_names:
-            self.theme = self._theme_definitions.get(themename)
-            # Creating the builder also runs theme_create + theme_use for the
-            # new theme and builds its default (".") styles.
-            self._theme_objects[themename] = StyleBuilderTTK()
-        else:
-            raise TclError(themename, "is not a valid theme.")
-
-        # Repaint the live widget tree for the new theme. Each theme has its
-        # own clam-derived Tcl style DB, so a style configured under one theme
-        # is invisible under another; the walk rebuilds the (theme, style)
-        # pairs that mounted widgets actually reference -- O(mounted), not
-        # O(all-styles-ever-used) -- and restyles legacy tk widgets inline.
-        self._theme_version += 1
-        self._theme_walk()
-
-    def theme_create(self, themename: str, parent: str = None, settings: dict = None) -> None:
-        """
-        Create a new theme in the Tcl interpreter. If the parent is a registered
-        ttkbootstrap theme, the new theme will be registered with a copied
-        ThemeDefinition and builder. Duplicate registration is avoided.
-
-        Parameters:
-
-            themename (str):
-                The name of the new theme.
-
-            parent (str):
-                The name of the parent theme to inherit from.
-
-            settings (dict):
-                A dictionary of style settings (Tcl-style).
-        """
-        from tkinter.ttk import _script_from_settings  # type: ignore[attr-defined]
-
-        script = _script_from_settings(settings) if settings else ''
-
-        # Lazy-load parent if it's a known bootstrap theme
-        if parent:
-            if parent not in super().theme_names():
-                if parent in self._theme_names:
-                    self.theme_use(parent)
-                else:
-                    raise TclError(f"{parent!r} is not a valid theme name or parent theme.")
-
-        # Create the Tcl-level theme
-        if parent:
-            self.tk.call(
-                self._name, "theme", "create", themename,
-                "-parent", parent, "-settings", script)
-        else:
-            self.tk.call(
-                self._name, "theme", "create", themename,
-                "-settings", script)
-
-        # Register the new theme if copying from a ttkbootstrap theme
-        if parent in self._theme_definitions and themename not in self._theme_definitions:
-            parent_def = self._theme_definitions[parent]
-            copied_def = ThemeDefinition(
-                name=themename,
-                colors=parent_def.colors,
-                themetype=parent_def.type
-            )
-            self._theme_definitions[themename] = copied_def
-            self._theme_names.add(themename)
-            self._theme_styles[themename] = set()
-
-            if themename not in self._theme_objects:
-                self._theme_objects[themename] = StyleBuilderTTK(build=False)
-
-    def style_exists_in_theme(self, ttkstyle: str):
-        """Check if a style exists in the current theme.
-
-        Parameters:
-
-            ttkstyle (str):
-                The ttk style to check.
-
-        Returns:
-
-            bool:
-                `True` if the style exists, otherwise `False`.
-        """
-        if self.theme is None:
-            return False
-
-        theme_styles = self._theme_styles.get(self.theme.name)
-        if theme_styles is None:
-            return False
-
-        exists_in_theme = ttkstyle in theme_styles
-        exists_in_registry = ttkstyle in self._style_registry
-        return exists_in_theme and exists_in_registry
-
-    def use_dynamic_foreground(self, enable: bool = True):
-        """Enable or disable dynamic foreground color selection.
-
-        When enabled, the foreground color of widgets will be decided
-        between the `fg` and `selectfg` colors based on the
-        contrast ratio with the widget's background color.
-        At default, this is disabled.
-
-        Parameters:
-
-            enable (bool):
-                If `True`, dynamic foreground selection is enabled.
-                Otherwise, it is disabled.
-        """
-        self._dynamic_foreground = enable
-
-    @property
-    def dynamic_foreground(self):
-        """Returns `True` if dynamic foreground selection is enabled,
-        otherwise `False`.
-        """
-        return self._dynamic_foreground
-
-    @staticmethod
-    def get_instance():
-        """Returns and instance of the style class"""
-        return Style.instance
-
-    @staticmethod
-    def _get_builder():
-        """Get the object that builds the widget styles for the current
-        theme.
-
-        Returns:
-
-            ThemeBuilderTTK:
-                The theme builder object that builds the ttk styles for
-                the current theme.
-        """
-        style: Style = Style.get_instance()
-        theme_name = style.theme.name
-        return style._theme_objects[theme_name]
-
-    @staticmethod
-    def _get_builder_tk():
-        """Get the object that builds the widget styles for the current
-        theme.
-
-        Returns:
-
-            ThemeBuilderTK:
-                The theme builder object that builds the ttk styles for
-                the current theme.
-        """
-        builder = Style._get_builder()
-        return builder.builder_tk
-
-    def _build_configure(self, style, **kw):
-        """Calls configure of superclass; used by style builder classes."""
-        super().configure(style, **kw)
-
-    def _load_themes(self, EXTERNAL_THEMES=None):
-        """Load all ttkbootstrap defined themes"""
-        # create a theme definition object for each theme, this will be
-        # used to generate the theme in tkinter along with any assets
-        # at run-time
-        if USER_THEMES:
-            STANDARD_THEMES.update(USER_THEMES)
-
-        if EXTERNAL_THEMES:
-            STANDARD_THEMES.update(EXTERNAL_THEMES)
-
-        theme_settings = {"themes": STANDARD_THEMES}
-        for name, definition in theme_settings["themes"].items():
-            self.register_theme(
-                ThemeDefinition(
-                    name=name,
-                    themetype=definition["type"],
-                    colors=definition["colors"],
-                )
-            )
-
-    def _register_ttkstyle(self, ttkstyle):
-        """Register that a ttk style name. This ensures that the
-        builder will not attempt to build a style that has already
-        been created.
-
-        Parameters:
-
-            ttkstyle (str):
-                The name of the ttk style to register.
-        """
-        self._style_registry.add(ttkstyle)
-        theme = self.theme.name
-        self._theme_styles[theme].add(ttkstyle)
-
-    def _theme_walk(self):
-        """Repaint the live widget tree for the current theme.
-
-        Depth-first over `winfo_children()` from the root, repainting only
-        widgets stamped older than `_theme_version` (so each widget is touched
-        at most once per switch). ttk widgets get their `(theme, style)`
-        rebuilt on demand; legacy tk widgets are restyled inline. Because the
-        walk reaches every mounted widget through the tree, it replaces both
-        the Publisher broadcast (legacy tk widgets) and the
-        rebuild-every-registered-style scan (ttk widgets).
-        """
-        root = self.master
-        if root is None:
-            return
-
-        version = self._theme_version
-        stack = [root]
-        while stack:
-            widget = stack.pop()
-            # Honor autostyle=False: such widgets opted out of theming and are
-            # never repainted, but their (autostyled) descendants still are.
-            if not getattr(widget, "_tb_no_autostyle", False):
-                if getattr(widget, "_theme_version", None) != version:
-                    self._repaint_widget(widget)
-                    try:
-                        widget._theme_version = version
-                    except (AttributeError, TypeError):
-                        # Builtin/extension widgets that forbid new attributes
-                        # are simply repainted on every switch -- correct, just
-                        # not skippable.
-                        pass
-            try:
-                stack.extend(widget.winfo_children())
-            except TclError:
-                pass
-
-    def _repaint_widget(self, widget):
-        """Restyle a single widget for the current theme.
-
-        ttk widgets ensure their style exists under the active theme (rebuilt
-        lazily by the builder if stale); legacy tk widgets re-run their tk
-        update method. The combobox popdown -- a Tcl toplevel the DFS cannot
-        reach -- is refreshed inside `update_ttk_widget_style`.
-        """
-        if isinstance(widget, ttk.Widget):
-            Bootstyle.update_ttk_widget_style(widget)
-        else:
-            Bootstyle.update_tk_widget_style(widget)
-
-    def _get_or_create_image(self, key, factory):
-        """Return the Tcl name of a cached asset image, building on a miss.
-
-        Content-addressed memoization: `key` must capture every
-        pixel-determining input of the asset (resolved colors, scaled pixel
-        size, geometry/variant) so identical assets dedupe across themes and
-        are rendered exactly once. The cache holds a strong reference to the
-        `PhotoImage`, keeping its Tcl image alive for as long as styles
-        reference it by name.
-
-        Parameters:
-
-            key:
-                A hashable tuple of the asset's pixel-determining inputs.
-
-            factory (Callable[[], ImageTk.PhotoImage]):
-                A zero-arg builder that renders the image. Called only on a
-                cache miss; must be pure with respect to `key`.
-
-        Returns:
-
-            str:
-                The tcl/tk image name to use in the style.
-        """
-        cached = self._image_cache.get(key)
-        if cached is not None:
-            return cached[0]
-        image = factory()
-        name = util.get_image_name(image)
-        self._image_cache[key] = (name, image)
-        return name
-
-    def clear_image_cache(self):
-        """Drop all cached widget asset images.
-
-        Releases the cached `PhotoImage` objects (and their underlying Tcl
-        images). This is a memory-reclamation and diagnostics escape hatch, not
-        a live refresh: styles are built once per `(theme, style)`, so styles
-        built before the clear keep referencing the freed image names until
-        they are rebuilt -- which happens when a not-yet-built theme is
-        activated, not on a same-theme switch. Clear when image widgets for the
-        affected styles are not currently displayed (e.g. before moving to a
-        fresh theme), or to reclaim memory at shutdown.
-        """
-        self._image_cache.clear()
-
-    def load_user_theme(self, theme: ThemeDefinition):
-        """Load a user theme definition"""
-        self.register_theme(theme)
-
-    def load_user_themes(self, file):
-        """Load user themes saved in json format"""
-        with open(file, encoding='utf-8') as f:
-            data = json.load(f)
-            themes = data['themes']
-        for theme in themes:
-            for name, definition in theme.items():
-                self.register_theme(
-                    ThemeDefinition(
-                        name=name,
-                        themetype=definition["type"],
-                        colors=definition["colors"],
-                    )
-                )
-
-
-class StyleBuilderTK:
-    """A class for styling legacy tkinter widgets (not ttk).
-
-    The methods in this classed are used internally to update tk widget
-    style configurations and are not intended to be called by the end
-    user.
-
-    All legacy tkinter widgets are updated with a callback whenever the
-    theme is changed. The color configuration of the widget is updated
-    to match the current theme. Legacy ttk widgets are not the primary
-    focus of this library, however, an attempt was made to make sure they
-    did not stick out amongst ttk widgets if used.
-
-    Some ttk widgets contain legacy components that must be updated
-    such as the Combobox popdown, so this ensures they are styled
-    completely to match the current theme.
-    """
-
-    def __init__(self):
-        self.style = Style.get_instance()
-        self.master = self.style.master
-
-    @property
-    def theme(self) -> ThemeDefinition:
-        """A reference to the `ThemeDefinition` object for the current
-        theme."""
-        return self.style.theme
-
-    @property
-    def colors(self) -> Colors:
-        """A reference to the `Colors` object for the current theme."""
-        return self.style.colors
-
-    @property
-    def is_light_theme(self) -> bool:
-        """Returns `True` if the theme is _light_, otherwise `False`."""
-        return self.style.theme.type == LIGHT
-
-    def update_tk_style(self, widget: tk.Tk):
-        """Update the window style.
-
-        Parameters:
-
-            widget (tkinter.Tk):
-                The tk object to update.
-        """
-        widget.configure(background=self.colors.bg)
-        # add default initial font for text widget
-        widget.option_add('*Text*Font', 'TkDefaultFont')
-
-    def update_toplevel_style(self, widget: tk.Toplevel):
-        """Update the toplevel style.
-
-        Parameters:
-
-            widget (tkinter.Toplevel):
-                The toplevel object to update.
-        """
-        widget.configure(background=self.colors.bg)
-
-    def update_canvas_style(self, widget: tk.Canvas):
-        """Update the canvas style.
-
-        Parameters:
-
-            widget (tkinter.Canvas):
-                The canvas object to update.
-        """
-        # if self.is_light_theme:
-        #     bordercolor = self.colors.border
-        # else:
-        #     bordercolor = self.colors.selectbg
-
-        widget.configure(
-            background=self.colors.bg,
-            highlightthickness=0,
-            # highlightbackground=bordercolor,
-        )
-
-    def update_button_style(self, widget: tk.Button):
-        """Update the button style.
-
-        Parameters:
-
-            widget (tkinter.Button):
-                The button object to update.
-        """
-        background = self.colors.primary
-        foreground = self.colors.selectfg
-        activebackground = Colors.update_hsv(self.colors.primary, vd=-0.1)
-
-        widget.configure(
-            background=background,
-            foreground=foreground,
-            relief=tk.FLAT,
-            borderwidth=0,
-            activebackground=activebackground,
-            highlightbackground=self.colors.selectfg,
-        )
-
-    def update_label_style(self, widget: tk.Label):
-        """Update the label style.
-
-        Parameters:
-
-            widget (tkinter.Label):
-                The label object to update.
-        """
-        widget.configure(foreground=self.colors.fg, background=self.colors.bg)
-
-    def update_frame_style(self, widget: tk.Frame):
-        """Update the frame style.
-
-        Parameters:
-
-            widget (tkinter.Frame):
-                The frame object to update.
-        """
-        widget.configure(background=self.colors.bg)
-
-    def update_checkbutton_style(self, widget: tk.Checkbutton):
-        """Update the checkbutton style.
-
-        Parameters:
-
-            widget (tkinter.Checkbutton):
-                The checkbutton object to update.
-        """
-        widget.configure(
-            activebackground=self.colors.bg,
-            activeforeground=self.colors.primary,
-            background=self.colors.bg,
-            foreground=self.colors.fg,
-            selectcolor=self.colors.bg,
-        )
-
-    def update_radiobutton_style(self, widget: tk.Radiobutton):
-        """Update the radiobutton style.
-
-        Parameters:
-
-            widget (tkinter.Radiobutton):
-                The radiobutton object to update.
-        """
-        widget.configure(
-            activebackground=self.colors.bg,
-            activeforeground=self.colors.primary,
-            background=self.colors.bg,
-            foreground=self.colors.fg,
-            selectcolor=self.colors.bg,
-        )
-
-    def update_entry_style(self, widget: tk.Entry):
-        """Update the entry style.
-
-        Parameters:
-
-            widget (tkinter.Entry):
-                The entry object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        widget.configure(
-            relief=tk.FLAT,
-            highlightthickness=1,
-            foreground=self.colors.inputfg,
-            highlightbackground=bordercolor,
-            highlightcolor=self.colors.primary,
-            background=self.colors.inputbg,
-            insertbackground=self.colors.inputfg,
-            insertwidth=1,
-        )
-
-    def update_scale_style(self, widget: tk.Scale):
-        """Update the scale style.
-
-        Parameters:
-
-            widget (tkinter.scale):
-                The scale object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        activecolor = Colors.update_hsv(self.colors.primary, vd=-0.2)
-        widget.configure(
-            background=self.colors.primary,
-            showvalue=False,
-            sliderrelief=tk.FLAT,
-            borderwidth=0,
-            activebackground=activecolor,
-            highlightthickness=1,
-            highlightcolor=bordercolor,
-            highlightbackground=bordercolor,
-            troughcolor=self.colors.inputbg,
-        )
-
-    def update_spinbox_style(self, widget: tk.Spinbox):
-        """Update the spinbox style.
-
-        Parameters:
-
-            widget (tkinter.Spinbox):
-                THe spinbox object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        widget.configure(
-            relief=tk.FLAT,
-            highlightthickness=1,
-            foreground=self.colors.inputfg,
-            highlightbackground=bordercolor,
-            highlightcolor=self.colors.primary,
-            background=self.colors.inputbg,
-            buttonbackground=self.colors.inputbg,
-            insertbackground=self.colors.inputfg,
-            insertwidth=1,
-            # these options should work, but do not have any affect
-            buttonuprelief=tk.FLAT,
-            buttondownrelief=tk.SUNKEN,
-        )
-
-    def update_listbox_style(self, widget: tk.Listbox):
-        """Update the listbox style.
-
-        Parameters:
-
-            widget (tkinter.Listbox):
-                The listbox object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        widget.configure(
-            foreground=self.colors.inputfg,
-            background=self.colors.inputbg,
-            selectbackground=self.colors.selectbg,
-            selectforeground=self.colors.selectfg,
-            highlightcolor=self.colors.primary,
-            highlightbackground=bordercolor,
-            highlightthickness=1,
-            activestyle="none",
-            relief=tk.FLAT,
-        )
-
-    def update_menubutton_style(self, widget: tk.Menubutton):
-        """Update the menubutton style.
-
-        Parameters:
-
-            widget (tkinter.Menubutton):
-                The menubutton object to update.
-        """
-        activebackground = Colors.update_hsv(self.colors.primary, vd=-0.2)
-        widget.configure(
-            background=self.colors.primary,
-            foreground=self.colors.selectfg,
-            activebackground=activebackground,
-            activeforeground=self.colors.selectfg,
-            borderwidth=0,
-        )
-
-    def update_menu_style(self, widget: tk.Menu):
-        """Update the menu style.
-
-        Parameters:
-
-            widget (tkinter.Menu):
-                The menu object to update.
-        """
-        widget.configure(
-            tearoff=False,
-            activebackground=self.colors.selectbg,
-            activeforeground=self.colors.selectfg,
-            foreground=self.colors.fg,
-            selectcolor=self.colors.primary,
-            background=self.colors.bg,
-            relief=tk.FLAT,
-            borderwidth=0,
-        )
-
-    def update_labelframe_style(self, widget: tk.LabelFrame):
-        """Update the labelframe style.
-
-        Parameters:
-
-            widget (tkinter.LabelFrame):
-                The labelframe object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        widget.configure(
-            highlightcolor=bordercolor,
-            foreground=self.colors.fg,
-            borderwidth=1,
-            highlightthickness=0,
-            background=self.colors.bg,
-        )
-
-    def update_text_style(self, widget: tk.Text):
-        """Update the text style.
-
-        Parameters:
-
-            widget (tkinter.Text):
-                The text object to update.
-        """
-        if self.is_light_theme:
-            bordercolor = self.colors.border
-        else:
-            bordercolor = self.colors.selectbg
-
-        focuscolor = widget.cget("highlightbackground")
-
-        if focuscolor in ["SystemButtonFace", bordercolor]:
-            focuscolor = bordercolor
-
-        widget.configure(
-            background=self.colors.inputbg,
-            foreground=self.colors.inputfg,
-            highlightcolor=focuscolor,
-            highlightbackground=bordercolor,
-            insertbackground=self.colors.inputfg,
-            selectbackground=self.colors.selectbg,
-            selectforeground=self.colors.selectfg,
-            insertwidth=1,
-            highlightthickness=1,
-            relief=tk.FLAT,
-            padx=5,
-            pady=5,
-            # font="TkDefaultFont",
-        )
+from ttkbootstrap.style.theme import Colors, ThemeDefinition
+from ttkbootstrap.style.builders_tk import StyleBuilderTK
 
 
 class StyleBuilderTTK:
@@ -1379,6 +26,9 @@ class StyleBuilderTTK:
     """
 
     def __init__(self, build: bool = True):
+        # local import breaks the builders<-engine cycle (engine imports builders)
+        from ttkbootstrap.style.engine import Style
+
         self.style: Style = Style.get_instance()
         self.builder_tk = StyleBuilderTK()
 
@@ -5051,820 +3701,3 @@ class StyleBuilderTTK:
         # set scrollbar style
         sb_style = "TCombobox.Vertical.TScrollbar"
         widget.tk.call(f"{popdown}.f.sb", "configure", "-style", sb_style)
-
-
-class Keywords:
-    """Static keyword lists and regex patterns used to parse
-    ttkbootstrap "bootstyle" strings.
-
-    Bootstyle strings contain space- or dash-separated tokens that may
-    specify: a widget class (e.g. "button"), an orientation ("horizontal"
-    or "vertical"), a color ("primary", "info", etc.), and optional
-    type modifiers (e.g. "outline", "link", "inverse", "striped"). The
-    constants and compiled regexes in this class centralize those token
-    definitions for reuse by the bootstyle parsing helpers.
-
-    This class is internal to the styling system and not intended to be
-    instantiated.
-    """
-    
-    COLORS = [
-        "primary",
-        "secondary",
-        "success",
-        "info",
-        "warning",
-        "danger",
-        "light",
-        "dark",
-    ]
-    ORIENTS = ["horizontal", "vertical"]
-    TYPES = [
-        "outline",
-        "link",
-        "inverse",
-        "round",
-        "square",
-        "striped",
-        "focus",
-        "input",
-        "date",
-        "metersubtxt",
-        "meter",
-        "table"
-    ]
-    CLASSES = [
-        "button",
-        "progressbar",
-        "checkbutton",
-        "combobox",
-        "entry",
-        "labelframe",
-        "label",
-        "frame",
-        "floodgauge",
-        "sizegrip",
-        "optionmenu",
-        "menubutton",
-        "menu",
-        "notebook",
-        "panedwindow",
-        "radiobutton",
-        "separator",
-        "scrollbar",
-        "spinbox",
-        "scale",
-        "text",
-        "toolbutton",
-        "treeview",
-        "toggle",
-        "tk",
-        "calendar",
-        "listbox",
-        "canvas",
-        "toplevel",
-    ]
-    COLOR_PATTERN = re.compile("|".join(COLORS))
-    ORIENT_PATTERN = re.compile("|".join(ORIENTS))
-    CLASS_PATTERN = re.compile("|".join(CLASSES))
-    TYPE_PATTERN = re.compile("|".join(TYPES))
-
-
-class Bootstyle:
-    """Helpers for parsing and applying ttkbootstrap "bootstyle" options.
-
-    Bootstyle augments ttk widgets with a compact styling API that lets
-    you configure color, orientation, and type with a single string (or
-    tuple) such as "primary-outline", "success", or ("danger", "inverse").
-
-    This class provides utilities to parse those tokens from strings and
-    widget state, determine the target widget class and orientation, and
-    resolve the requested color and variant. Its `update_ttk_widget_style`
-    resolver is the engine shared by both api-delivery paths: the default
-    `BootMixin` subclasses and the opt-in global patch (`enable_global_api`).
-
-    Typical end users will not call these methods directly; they are used
-    internally by the Style engine, the mixins, and the global patch.
-    """
-    @staticmethod
-    def ttkstyle_widget_class(widget=None, string=""):
-        """Find and return the widget class
-
-        Parameters:
-
-            widget (Widget):
-                The widget object.
-
-            string (str):
-                A keyword string to parse.
-
-        Returns:
-
-            str:
-                A widget class keyword.
-        """
-        # find widget class from string pattern
-        match = re.search(Keywords.CLASS_PATTERN, string.lower())
-        if match is not None:
-            widget_class = match.group(0)
-            return widget_class
-
-        # find widget class from tkinter/tcl method
-        if widget is None:
-            return ""
-        _class = widget.winfo_class()
-        match = re.search(Keywords.CLASS_PATTERN, _class.lower())
-        if match is not None:
-            widget_class = match.group(0)
-            return widget_class
-        else:
-            return ""
-
-    @staticmethod
-    def ttkstyle_widget_type(string):
-        """Find and return the widget type.
-
-        Parameters:
-
-            string (str):
-                A keyword string to parse.
-
-        Returns:
-
-            str:
-                A widget type keyword.
-        """
-        match = re.search(Keywords.TYPE_PATTERN, string.lower())
-        if match is None:
-            return ""
-        else:
-            widget_type = match.group(0)
-            return widget_type
-
-    @staticmethod
-    def ttkstyle_widget_orient(widget=None, string="", **kwargs):
-        """Find and return widget orient, or default orient for widget if
-        a widget with orientation.
-
-        Parameters:
-
-            widget (Widget):
-                The widget object.
-
-            string (str):
-                A keyword string to parse.
-
-            **kwargs:
-                Optional keyword arguments passed in the widget constructor.
-
-        Returns:
-
-            str:
-                A widget orientation keyword.
-        """
-        # string method (priority)
-        match = re.search(Keywords.ORIENT_PATTERN, string)
-        widget_orient = ""
-
-        if match is not None:
-            widget_orient = match.group(0)
-            return widget_orient
-
-        # orient from kwargs
-        if "orient" in kwargs:
-            _orient = kwargs.pop("orient")
-            if _orient == "h":
-                widget_orient = "horizontal"
-            elif _orient == "v":
-                widget_orient = "vertical"
-            else:
-                widget_orient = _orient
-            return widget_orient
-
-        # orient from settings
-        if widget is None:
-            return widget_orient
-        try:
-            widget_orient = str(widget.cget("orient"))
-        except:
-            pass
-
-        return widget_orient
-
-    @staticmethod
-    def ttkstyle_widget_color(string):
-        """Find and return widget color
-
-        Parameters:
-
-            string (str):
-                A keyword string to parse.
-
-        Returns:
-
-            str:
-                A color keyword.
-        """
-        _color = re.search(Keywords.COLOR_PATTERN, string.lower())
-        if _color is None:
-            return ""
-        else:
-            widget_color = _color.group(0)
-            return widget_color
-
-    @staticmethod
-    def ttkstyle_name(widget=None, string="", **kwargs):
-        """Parse a string to build and return a ttkstyle name.
-
-        Parameters:
-
-            widget (Widget):
-                The widget object.
-
-            string (str):
-                A keyword string to parse.
-
-            **kwargs:
-                Other keyword arguments to parse widget orientation.
-
-        Returns:
-
-            str:
-                A ttk style name
-        """
-        style_string = "".join(string).lower()
-        widget_color = Bootstyle.ttkstyle_widget_color(style_string)
-        widget_type = Bootstyle.ttkstyle_widget_type(style_string)
-        widget_orient = Bootstyle.ttkstyle_widget_orient(
-            widget, style_string, **kwargs
-        )
-        widget_class = Bootstyle.ttkstyle_widget_class(widget, style_string)
-
-        if widget_color:
-            widget_color = f"{widget_color}."
-
-        if widget_type:
-            widget_type = f"{widget_type.title()}."
-
-        if widget_orient:
-            widget_orient = f"{widget_orient.title()}."
-
-        if widget_class.startswith("t"):
-            widget_class = widget_class.title()
-        else:
-            widget_class = f"T{widget_class.title()}"
-
-        ttkstyle = f"{widget_color}{widget_type}{widget_orient}{widget_class}"
-        return ttkstyle
-
-    @staticmethod
-    def ttkstyle_method_name(widget=None, string=""):
-        """Parse a string to build and return the name of the
-        `StyleBuilderTTK` method that creates the ttk style for the
-        target widget.
-
-        Parameters:
-
-            widget (Widget):
-                The widget object to lookup.
-
-            string (str):
-                The keyword string to parse.
-
-        Returns:
-
-            str:
-                The name of the update method used to update the widget.
-        """
-        style_string = "".join(string).lower()
-        widget_type = Bootstyle.ttkstyle_widget_type(style_string)
-        widget_class = Bootstyle.ttkstyle_widget_class(widget, style_string)
-
-        if widget_type:
-            widget_type = f"_{widget_type}"
-
-        if widget_class:
-            widget_class = f"_{widget_class}"
-
-        if not widget_type and not widget_class:
-            return ""
-        else:
-            method_name = f"create{widget_type}{widget_class}_style"
-            return method_name
-
-    @staticmethod
-    def tkupdate_method_name(widget):
-        """Lookup the tkinter style update method from the widget class
-
-        Parameters:
-
-            widget (Widget):
-                The widget object to lookup.
-
-        Returns:
-
-            str:
-                The name of the method used to update the widget object.
-        """
-        widget_class = Bootstyle.ttkstyle_widget_class(widget)
-
-        if widget_class:
-            widget_class = f"_{widget_class}"
-
-        method_name = f"update{widget_class}_style"
-        return method_name
-
-    @staticmethod
-    def override_ttk_widget_constructor(func):
-        """Override widget constructors with bootstyle api options.
-
-        Parameters:
-
-            func (Callable):
-                The widget class `__init__` method
-        """
-
-        def __init__(self, *args, **kwargs):
-
-            # Blessed subclasses (BootMixin) own bootstyle resolution; when the
-            # global API is also enabled, defer to the mixin to avoid
-            # double-applying a style.
-            if isinstance(self, BootMixin):
-                func(self, *args, **kwargs)
-                return
-
-            # capture bootstyle and style arguments
-            if "bootstyle" in kwargs:
-                bootstyle = kwargs.pop("bootstyle")
-            else:
-                bootstyle = ""
-
-            if "style" in kwargs:
-                style = kwargs.pop("style") or ""
-            else:
-                style = ""
-
-            # instantiate the widget
-            func(self, *args, **kwargs)
-
-            # must be called AFTER instantiation in order to use winfo_class
-            #    in the `get_ttkstyle_name` method
-
-            try:
-                if style:
-                    if Style.get_instance().style_exists_in_theme(style):
-                        self.configure(style=style)
-                    else:
-                        ttkstyle = Bootstyle.update_ttk_widget_style(
-                            self, style, **kwargs
-                        )
-                        self.configure(style=ttkstyle)
-                elif bootstyle:
-                    ttkstyle = Bootstyle.update_ttk_widget_style(
-                        self, bootstyle, **kwargs
-                    )
-                    self.configure(style=ttkstyle)
-                else:
-                    ttkstyle = Bootstyle.update_ttk_widget_style(
-                        self, "default", **kwargs
-                    )
-                    self.configure(style=ttkstyle)
-                # Stamp the widget current so the next theme walk only repaints
-                # it once the theme actually changes.
-                Bootstyle.stamp_theme_version(self)
-            except AttributeError:
-                # Third-party widgets (e.g. tkcalendar.Calendar) override
-                # configure() and may access instance attributes that are not
-                # yet set when ttk.Frame.__init__ calls back into the
-                # subclass configure before the subclass __init__ completes.
-                pass
-
-        return __init__
-
-    @staticmethod
-    def override_ttk_widget_configure(func):
-        """Overrides the configure method on a ttk widget.
-
-        Parameters:
-
-            func (Callable):
-                The widget class `configure` method
-        """
-
-        def configure(self, cnf=None, **kwargs):
-            # Blessed subclasses (BootMixin) own bootstyle resolution; defer to
-            # the mixin's configure when the global API is also enabled.
-            if isinstance(self, BootMixin):
-                return func(self, cnf, **kwargs)
-
-            # get configuration
-            if cnf in ("bootstyle", "style"):
-                return self.cget("style")
-
-            if cnf is not None:
-                return func(self, cnf)
-
-            # set configuration
-            if "bootstyle" in kwargs:
-                bootstyle = kwargs.pop("bootstyle")
-            else:
-                bootstyle = ""
-
-            if "style" in kwargs:
-                style = kwargs.get("style")
-                ttkstyle = Bootstyle.update_ttk_widget_style(
-                    self, style, **kwargs
-                )
-            elif bootstyle:
-                ttkstyle = Bootstyle.update_ttk_widget_style(
-                    self, bootstyle, **kwargs
-                )
-                kwargs.update(style=ttkstyle)
-
-            # update widget configuration
-            func(self, cnf, **kwargs)
-
-        return configure
-
-    @staticmethod
-    def update_ttk_widget_style(
-            widget: ttk.Widget = None, style_string: str = None, **kwargs
-    ):
-        """Update the ttk style or create if not existing.
-
-        Parameters:
-
-            widget (ttk.Widget):
-                The widget instance being updated.
-
-            style_string (str):
-                The style string to evalulate. May be the `style`, `ttkstyle`
-                or `bootstyle` argument depending on the context and scenario.
-
-            **kwargs:
-                Other optional keyword arguments.
-
-        Returns:
-
-            str:
-                The ttkstyle or empty string if there is none.
-        """
-        style: Style = Style.get_instance() or Style()
-
-        # get existing widget style if not provided
-        if style_string is None:
-            style_string = widget.cget("style")
-
-        # do nothing if the style has not been set
-        if not style_string:
-            return ""
-
-        if style_string == '.':
-            return '.'
-
-        # build style if not existing (example: theme changed)
-        ttkstyle = Bootstyle.ttkstyle_name(widget, style_string, **kwargs)
-        if not style.style_exists_in_theme(ttkstyle):
-            widget_color = Bootstyle.ttkstyle_widget_color(ttkstyle)
-            method_name = Bootstyle.ttkstyle_method_name(widget, ttkstyle)
-            builder: StyleBuilderTTK = style._get_builder()
-            try:
-                builder_method = builder.name_to_method(method_name)
-            except AttributeError:
-                # Style name is from a third-party widget (e.g. tkcalendar) and
-                # doesn't map to any ttkbootstrap builder; pass it through as-is.
-                return style_string
-            builder_method(builder, widget_color)
-
-        # Repaint the combobox popdown. It is a Tcl-level toplevel that the
-        # theme walk's winfo_children() DFS cannot reach, so it is refreshed
-        # here instead -- the walk calls this method for every ttk widget on a
-        # theme change, which keeps the popdown in sync without a subscription.
-        try:
-            if widget.winfo_class() == "TCombobox":
-                builder: StyleBuilderTTK = style._get_builder()
-                builder.update_combobox_popdown_style(widget)
-        except (AttributeError, TclError):
-            pass
-
-        return ttkstyle
-
-    @staticmethod
-    def setup_ttkbootstrap_api():
-        """Monkey-patch the stock tkinter/ttk widget classes with the
-        bootstyle/autostyle api.
-
-        As of 2.0 this is the legacy *global* path and is no longer run at
-        import time; the default api is delivered through concrete subclasses
-        (`BootMixin`/`AutoStyleMixin`). Call `enable_global_api` to opt back
-        into patching the stock classes (e.g. for code that creates vanilla
-        ``tkinter.ttk`` widgets)."""
-        from ttkbootstrap.widgets import TTK_WIDGETS
-        from ttkbootstrap.widgets import TK_WIDGETS
-
-        # TTK WIDGETS
-        for widget in TTK_WIDGETS:
-            try:
-                # override widget constructor
-                _init = Bootstyle.override_ttk_widget_constructor(
-                    widget.__init__
-                )
-                widget.__init__ = _init
-
-                # override configure method
-                _configure = Bootstyle.override_ttk_widget_configure(
-                    widget.configure
-                )
-                widget.configure = _configure
-                widget.config = widget.configure
-
-                # override get and set methods
-                _orig_getitem = widget.__getitem
-                _orig_setitem = widget.__setitem
-
-                def __setitem(self, key, val):
-                    if key in ("bootstyle", "style"):
-                        return _configure(self, **{key: val})
-                    return _orig_setitem(key, val)
-
-                def __getitem(self, key):
-                    if key in ("bootstyle", "style"):
-                        return _configure(self, cnf=key)
-                    return _orig_getitem(key)
-
-                if (
-                        widget.__name__ != "OptionMenu"
-                ):  # this has it's own override
-                    widget.__setitem__ = __setitem
-                    widget.__getitem__ = __getitem
-            except:
-                # this may fail in python 3.6 for ttk widgets that do not exist
-                #   in that version.
-                continue
-
-        # TK WIDGETS
-        for widget in TK_WIDGETS:
-            # override widget constructor
-            _init = Bootstyle.override_tk_widget_constructor(widget.__init__)
-            widget.__init__ = _init
-
-    @staticmethod
-    def stamp_theme_version(widget):
-        """Stamp a widget with the current theme version.
-
-        Freshly built widgets are already painted for the active theme, so
-        stamping them lets the next theme walk skip them until the theme
-        actually changes. No-ops cleanly before a `Style` exists or for
-        widgets that forbid new attributes.
-        """
-        style = Style.get_instance()
-        if style is None:
-            return
-        try:
-            widget._theme_version = style._theme_version
-        except (AttributeError, TypeError):
-            pass
-
-    @staticmethod
-    def update_tk_widget_style(widget):
-        """Lookup the widget name and call the appropriate update
-        method
-
-        Parameters:
-
-            widget (object):
-                The tcl/tk name given by `tkinter.Widget.winfo_name()`
-        """
-        try:
-            style = Style.get_instance()
-            method_name = Bootstyle.tkupdate_method_name(widget)
-            builder = style._get_builder_tk()
-            builder_method = getattr(StyleBuilderTK, method_name)
-            builder_method(builder, widget)
-        except:
-            """Must pass here to prevent a failure when the user calls
-            the `Style`method BEFORE an instance of `Tk` is instantiated.
-            This will defer the update of the `Tk` background until the end
-            of the `BootStyle` object instantiation (created by the `Style`
-            method)"""
-            pass
-
-    @staticmethod
-    def override_tk_widget_constructor(func):
-        """Override widget constructors to apply default style for tk
-        widgets.
-
-        Parameters:
-
-            func (Callable):
-                The `__init__` method for this widget.
-        """
-
-        def __init__wrapper(self, *args, **kwargs):
-
-            # Blessed subclasses (AutoStyleMixin) own autostyle handling; defer
-            # to the mixin when the global API is also enabled.
-            if isinstance(self, AutoStyleMixin):
-                func(self, *args, **kwargs)
-                return
-
-            # check for autostyle flag
-            if "autostyle" in kwargs:
-                autostyle = kwargs.pop("autostyle")
-            else:
-                autostyle = True
-
-            # instantiate the widget
-            func(self, *args, **kwargs)
-
-            if autostyle:
-                Bootstyle.update_tk_widget_style(self)
-                # Stamp the widget current so the next theme walk only
-                # repaints it once the theme actually changes.
-                Bootstyle.stamp_theme_version(self)
-            else:
-                # Opt out of theming entirely: the theme walk must skip this
-                # widget on switch, matching the pre-2.0 behavior where an
-                # unstyled widget never subscribed for repaints.
-                self._tb_no_autostyle = True
-
-        return __init__wrapper
-
-
-class BootMixin:
-    """Mixin that adds the ``bootstyle`` API to a ttk widget class.
-
-    This is the 2.0 delivery vehicle for the styling API. Instead of
-    monkey-patching ttk widget classes at import time, ttkbootstrap ships
-    concrete subclasses such as ``class Button(BootMixin, ttk.Button)`` and
-    re-exports them. Mixing this in front of any ttk-derived class gives that
-    class:
-
-    - a ``bootstyle=`` (and bare ``style=``) keyword on the constructor,
-    - ``configure``/``config`` that accept and report ``bootstyle``,
-    - ``widget["bootstyle"]`` / ``widget["bootstyle"] = ...`` access.
-
-    All resolution flows through the unchanged `Bootstyle.update_ttk_widget_style`
-    engine, so the mixin only changes *delivery* — not how a style string maps
-    to a ttk style. Because the accessors are real methods that use ``super()``
-    (not closures captured in a loop), the late-binding bug of the old
-    monkey-patch is gone.
-    """
-
-    def __init__(self, *args, **kwargs):
-        # capture bootstyle and style arguments
-        bootstyle = kwargs.pop("bootstyle", "")
-        style = kwargs.pop("style", "") or ""
-
-        # instantiate the underlying ttk widget first so winfo_class() is
-        # available to the resolver below
-        super().__init__(*args, **kwargs)
-
-        try:
-            if style:
-                if Style.get_instance().style_exists_in_theme(style):
-                    super().configure(style=style)
-                else:
-                    ttkstyle = Bootstyle.update_ttk_widget_style(
-                        self, style, **kwargs
-                    )
-                    super().configure(style=ttkstyle)
-            elif bootstyle:
-                ttkstyle = Bootstyle.update_ttk_widget_style(
-                    self, bootstyle, **kwargs
-                )
-                super().configure(style=ttkstyle)
-            else:
-                ttkstyle = Bootstyle.update_ttk_widget_style(
-                    self, "default", **kwargs
-                )
-                super().configure(style=ttkstyle)
-            # Stamp the widget current so the next theme walk only repaints it
-            # once the theme actually changes.
-            Bootstyle.stamp_theme_version(self)
-        except AttributeError:
-            # Third-party widgets (e.g. tkcalendar.Calendar) override
-            # configure() and may touch instance attributes not yet set when
-            # ttk.Frame.__init__ calls back into the subclass configure before
-            # the subclass __init__ completes.
-            pass
-
-    def configure(self, cnf=None, **kwargs):
-        # query a single option
-        if cnf in ("bootstyle", "style"):
-            return self.cget("style")
-        if cnf is not None:
-            return super().configure(cnf)
-
-        # set configuration
-        bootstyle = kwargs.pop("bootstyle", "")
-        if "style" in kwargs:
-            style = kwargs.get("style")
-            Bootstyle.update_ttk_widget_style(self, style, **kwargs)
-        elif bootstyle:
-            ttkstyle = Bootstyle.update_ttk_widget_style(
-                self, bootstyle, **kwargs
-            )
-            kwargs.update(style=ttkstyle)
-
-        return super().configure(cnf, **kwargs)
-
-    config = configure
-
-    def __setitem__(self, key, value):
-        if key in ("bootstyle", "style"):
-            return self.configure(**{key: value})
-        return super().__setitem__(key, value)
-
-    def __getitem__(self, key):
-        if key in ("bootstyle", "style"):
-            return self.cget("style")
-        return super().__getitem__(key)
-
-
-class AutoStyleMixin:
-    """Mixin that auto-applies the theme to a legacy ``tk`` widget class.
-
-    The tk counterpart to `BootMixin`. Legacy tk widgets have no ttk style;
-    instead they are painted with the active theme's colors at construction
-    via `Bootstyle.update_tk_widget_style`. Concrete subclasses such as
-    ``class Canvas(AutoStyleMixin, tk.Canvas)`` are re-exported from
-    ttkbootstrap.
-
-    Passing ``autostyle=False`` opts the widget out of theming entirely: it
-    keeps its native tk look and the theme walk skips it on switch (the
-    ``_tb_no_autostyle`` flag), matching the pre-2.0 behavior where an
-    unstyled widget never subscribed for repaints.
-    """
-
-    def __init__(self, *args, **kwargs):
-        autostyle = kwargs.pop("autostyle", True)
-        super().__init__(*args, **kwargs)
-        if autostyle:
-            Bootstyle.update_tk_widget_style(self)
-            Bootstyle.stamp_theme_version(self)
-        else:
-            self._tb_no_autostyle = True
-
-
-def bootify(cls):
-    """Return a ``bootstyle``-enabled subclass of any ttk widget class.
-
-    Use this to wrap a third-party ttk-derived widget that ttkbootstrap does
-    not ship a blessed subclass for:
-
-    ```python
-    ThemedCalendar = bootify(tkcalendar.Calendar)
-    cal = ThemedCalendar(root, bootstyle="info")
-    ```
-
-    The result is ``type(cls.__name__, (BootMixin, cls), {})`` — i.e. the same
-    construction used for the built-in blessed widgets, so it gains the full
-    ``bootstyle`` API without mutating the original class.
-    """
-    return type(cls.__name__, (BootMixin, cls), {})
-
-
-def apply_bootstyle(widget, bootstyle):
-    """Apply a bootstyle to an existing widget instance, no class mutation.
-
-    For per-instance styling of a widget whose class was never wrapped (for
-    example a plain ``tkinter.ttk`` widget): resolves the bootstyle to a ttk
-    style, assigns it, and stamps the widget current for the theme walk.
-
-    ```python
-    from tkinter import ttk
-    b = ttk.Button(root, text="Save")
-    apply_bootstyle(b, "success")
-    ```
-
-    Returns the resolved ttk style name.
-    """
-    ttkstyle = Bootstyle.update_ttk_widget_style(widget, bootstyle)
-    widget.configure(style=ttkstyle)
-    Bootstyle.stamp_theme_version(widget)
-    return ttkstyle
-
-
-_global_api_installed = False
-
-
-def enable_global_api():
-    """Re-apply the legacy global monkey-patch (opt-in).
-
-    In 2.0 the default styling API is delivered through concrete subclasses
-    (`BootMixin`/`AutoStyleMixin`), so importing ttkbootstrap no longer mutates
-    the stock ``tkinter``/``tkinter.ttk`` classes. Call this once if you have
-    code that creates *vanilla* ttk/tk widgets (e.g. ``from tkinter import
-    ttk; ttk.Button(..., bootstyle=...)``) and want the ``bootstyle``/
-    ``autostyle`` keywords on them anyway.
-
-    Idempotent. The installed wrappers defer to `BootMixin`/`AutoStyleMixin`
-    instances (the blessed subclasses), so enabling the global API never
-    double-resolves a style for a widget that already carries a mixin.
-    """
-    global _global_api_installed
-    if _global_api_installed:
-        return
-    Bootstyle.setup_ttkbootstrap_api()
-    _global_api_installed = True
